(chore) O3-5695: Unify encounter parsing across current-visit and past-visit
