+++ packages/esm-service-queues-app/src/current-visit/visit-details/current-visit-details.component.tsx
@@ -1,9 +1,9 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { StructuredListBody, StructuredListCell, StructuredListRow, StructuredListWrapper } from '@carbon/react';
-import { type OpenmrsResource, type Visit, formatTime, parseDate, useConfig } from '@openmrs/esm-framework';
-import { type ConfigObject } from '../../config-schema';
-import { type Note, type Encounter, type Observation, type DiagnosisItem } from '../../types/index';
+import { type OpenmrsResource, type Visit } from '@openmrs/esm-framework';
+import { type Encounter } from '../../types/index';
+import { useNotesAndDiagnoses } from '../hooks/useNotesAndDiagnoses';
 import { useVitalsFromObs } from '../hooks/useVitalsConceptMetadata';
 import VisitNote from './visit-note.component';
 import Vitals from './vitals.component';
@@ -22,47 +22,7 @@ enum visitTypes {
 
 const CurrentVisitDetails: React.FC<CurrentVisitProps> = ({ patientUuid, encounters, visit }) => {
   const { t } = useTranslation();
-  const { concepts, visitNoteEncounterTypeUuid } = useConfig<ConfigObject>();
-
-  const [diagnoses, notes, vitalsToRetrieve]: [Array<DiagnosisItem>, Array<Note>, Array<Encounter>] = useMemo(() => {
-    const notes: Array<Note> = [];
-    const vitalsToRetrieve: Array<Encounter> = [];
-    const diagnoses: Array<DiagnosisItem> = [];
-
-    // Iterating through every Encounter
-    encounters?.forEach((enc: Encounter) => {
-      // Gate on encounter type so notes/diagnoses concepts reused elsewhere don't leak in here.
-      if (enc.encounterType?.uuid === visitNoteEncounterTypeUuid) {
-        enc.obs?.forEach((obs: Observation) => {
-          if (obs.concept?.uuid === concepts.visitDiagnosesConceptUuid) {
-            const problemList = obs.groupMembers?.find((mem) => mem.concept?.uuid === concepts.problemListConceptUuid);
-            if (problemList?.value?.display) {
-              diagnoses.push({ diagnosis: problemList.value.display });
-            }
-          } else if (obs.concept?.uuid === concepts.generalPatientNoteConceptUuid) {
-            notes.push({
-              note: obs.value,
-              provider: {
-                name: enc.encounterProviders.length ? enc.encounterProviders[0].provider.person.display : '',
-                role: enc.encounterProviders.length ? enc.encounterProviders[0].encounterRole.display : '',
-              },
-              time: formatTime(parseDate(obs.obsDatetime)),
-              concept: obs.concept,
-            });
-          }
-        });
-      }
-
-      vitalsToRetrieve.push(enc);
-    });
-    return [diagnoses, notes, vitalsToRetrieve];
-  }, [
-    encounters,
-    visitNoteEncounterTypeUuid,
-    concepts.generalPatientNoteConceptUuid,
-    concepts.problemListConceptUuid,
-    concepts.visitDiagnosesConceptUuid,
-  ]);
+  const { notes, diagnoses } = useNotesAndDiagnoses(encounters as Array<Encounter>);
 
   return (
     <div className={styles.wrapper}>
@@ -80,7 +40,7 @@ const CurrentVisitDetails: React.FC<CurrentVisitProps> = ({ patientUuid, encount
               <StructuredListCell>
                 {' '}
                 <Vitals
-                  vitals={useVitalsFromObs(vitalsToRetrieve)}
+                  vitals={useVitalsFromObs(encounters)}
                   patientUuid={patientUuid}
                   visitType={visitTypes.CURRENT}
                   visit={visit}
